+++ R/conditions.R
@@ -2,12 +2,14 @@
                             class = "recoverome_error_input",
                             component = NULL,
                             ids = character(),
-                            call = rlang::caller_env()) {
-  rlang::abort(
+                            call = parent.frame(),
+                            .envir = parent.frame()) {
+  cli::cli_abort(
     message,
     class = c(class, "recoverome_error"),
     component = component,
     ids = ids,
-    call = call
+    call = call,
+    .envir = .envir
   )
 }